extract seed code into its own file
